+++ components/account/AccountPage.tsx
@@ -77,9 +77,6 @@ const AccountPage = (props: any) => {
         <AccountTransactions />
       </Grid>
 
-      <Grid item xs={12} md={8} className={classes.bottomPadding}>
-        <InternalTransactions />
-      </Grid>
 
       <Grid item xs={12} md={8} className={classes.bottomPadding}>
         <CoinBalanceHistory />